Add ErrorHandler to AbstractTestUI
